done fix code

+++ src/utils/CurrencyFormatterUtils.java
@@ -1,11 +1,12 @@
 package utils;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 import java.util.Locale;
 
 public class CurrencyFormatterUtils {
-    public static String formatCurrency(double amount) {
+    public static String formatCurrency(BigDecimal amount) {
         DecimalFormatSymbols symbols = new DecimalFormatSymbols(new Locale("vi", "VN"));
         symbols.setGroupingSeparator('.');
         DecimalFormat formatter = new DecimalFormat("#,###", symbols);
